Some reorganization and cleanup
