+++ src/core/mod.rs
@@ -1,0 +1,7 @@
+//! core components and types used throughout this library
+
+/// Error types used throughout this library.
+pub mod errors;
+
+/// Counter type defining operations required by the histogram and impls for primitives.
+pub mod counter;
