+++ src/lib.rs
@@ -212,48 +212,6 @@ const ORIGINAL_MIN: u64 = (-1_i64 >> 63) as u64;
 /// Max value of a new histogram.
 const ORIGINAL_MAX: u64 = 0;
 
-/// This trait represents the operations a histogram must be able to perform on the underlying
-/// counter type. The `ToPrimitive` trait is needed to perform floating point operations on the
-/// counts (usually for quantiles). The `FromPrimitive` to convert back into an integer count.
-/// Partial ordering is used for threshholding, also usually in the context of quantiles.
-pub trait Counter
-    : num::Num + num::ToPrimitive + num::FromPrimitive + num::Saturating + num::CheckedSub
-    + num::CheckedAdd + Copy + PartialOrd<Self> {
-
-    /// Counter as a f64.
-    fn as_f64(&self) -> f64;
-    /// Counter as a u64.
-    fn as_u64(&self) -> u64;
-}
-
-impl Counter for u8 {
-    #[inline]
-    fn as_f64(&self) -> f64 { *self as f64 }
-    #[inline]
-    fn as_u64(&self) -> u64 { *self as u64 }
-}
-
-impl Counter for u16 {
-    #[inline]
-    fn as_f64(&self) -> f64 { *self as f64 }
-    #[inline]
-    fn as_u64(&self) -> u64 { *self as u64 }
-}
-
-impl Counter for u32 {
-    #[inline]
-    fn as_f64(&self) -> f64 { *self as f64 }
-    #[inline]
-    fn as_u64(&self) -> u64 { *self as u64 }
-}
-
-impl Counter for u64 {
-    #[inline]
-    fn as_f64(&self) -> f64 { *self as f64 }
-    #[inline]
-    fn as_u64(&self) -> u64 { *self }
-}
-
 /// `Histogram` is the core data structure in HdrSample. It records values, and performs analytics.
 ///
 /// At its heart, it keeps the count for recorded samples in "buckets" of values. The resolution
@@ -325,76 +283,6 @@ pub struct Histogram<T: Counter> {
     total_count: u64,
     counts: Vec<T>,
 }
-
-/// Errors that can occur when creating a histogram.
-#[derive(Debug, Eq, PartialEq, Clone, Copy)]
-pub enum CreationError {
-    /// Lowest discernible value must be >= 1.
-    LowIsZero,
-    /// Lowest discernible value must be <= `u64::max_value() / 2` because the highest value is
-    /// a `u64` and the lowest value must be no bigger than half the highest.
-    LowExceedsMax,
-    /// Highest trackable value must be >= 2 * lowest discernible value for some internal
-    /// calculations to work out. In practice, high is typically much higher than 2 * low.
-    HighLessThanTwiceLow,
-    /// Number of significant digits must be in the range `[0, 5]`. It is capped at 5 because 5
-    /// significant digits is already more than almost anyone needs, and memory usage scales
-    /// exponentially as this increases.
-    SigFigExceedsMax,
-    /// Cannot represent sigfig worth of values beyond the lowest discernible value. Decrease the
-    /// significant figures, lowest discernible value, or both.
-    ///
-    /// This could happen if low is very large (like 2^60) and sigfigs is 5, which requires 18
-    /// additional bits, which would then require more bits than will fit in a u64. Specifically,
-    /// the exponent of the largest power of two that is smaller than the lowest value and the bits
-    /// needed to represent the requested significant figures must sum to 63 or less.
-    CannotRepresentSigFigBeyondLow,
-    /// The `usize` type is too small to represent the desired configuration. Use fewer significant
-    /// figures or a lower max.
-    UsizeTypeTooSmall
-}
-
-// TODO like RecordError, this is also an awkward split along resizing.
-/// Errors that can occur when adding another histogram.
-#[derive(Debug, Eq, PartialEq, Clone, Copy)]
-pub enum AdditionError {
-    /// The other histogram includes values that do not fit in this histogram's range.
-    /// Only possible when auto resize is disabled.
-    OtherAddendValueExceedsRange,
-    /// The other histogram includes values that would map to indexes in this histogram that are
-    /// not expressible for `usize`. Configure this histogram to use fewer significant digits. Only
-    /// possible when resize is enabled.
-    ResizeFailedUsizeTypeTooSmall
-}
-
-/// Errors that can occur when subtracting another histogram.
-#[derive(Debug, Eq, PartialEq, Clone, Copy)]
-pub enum SubtractionError {
-    /// The other histogram includes values that do not fit in this histogram's range.
-    /// Only possible when auto resize is disabled.
-    SubtrahendValueExceedsMinuendRange,
-    /// The other histogram includes counts that are higher than the current count for a value, and
-    /// counts cannot go negative. The subtraction may have been partially applied to some counts as
-    /// this error is returned when the first impossible subtraction is detected.
-    SubtrahendCountExceedsMinuendCount
-}
-
-// TODO the error conditions here are awkward: one only possible when resize is disabled, the other
-// only when resize is enabled.
-/// Errors that can occur while recording a value and its associated count.
-#[derive(Debug, Eq, PartialEq, Clone, Copy)]
-pub enum RecordError {
-    /// The value to record is not representable in this histogram and resizing is disabled.
-    /// Configure a higher maximum value or enable resizing. Only possible when resizing is
-    /// disabled.
-    ValueOutOfRangeResizeDisabled,
-    /// Auto resizing is enabled and must be used to represent the provided value, but the histogram
-    /// cannot be resized because `usize` cannot represent sufficient length. Configure this
-    /// histogram to use fewer significant digits. Only possible when resizing is enabled.
-    ResizeFailedUsizeTypeTooSmall
-}
-
-struct UsizeTypeTooSmall;
 
 /// Module containing the implementations of all `Histogram` iterators.
 pub mod iterators;
@@ -1721,5 +1609,7 @@ impl<T: Counter, F: Counter> PartialEq<Histogram<F>> for Histogram<T>
 mod tests;
 
 #[cfg(feature = "serialization")]
-#[path = "serialization/serialization.rs"]
 pub mod serialization;
+mod core;
+pub use core::errors::*;
+pub use core::counter::*;
